+++ vite.settings.config.ts
@@ -2,10 +2,9 @@ import path from 'node:path'
 import { defineConfig } from 'vite'
 import renderer from 'vite-plugin-electron-renderer'
 import vue from '@vitejs/plugin-vue'
-import cssInjectedByJsPlugin from 'vite-plugin-css-injected-by-js'
 
 export default defineConfig({
-  plugins: [ renderer(), vue(), cssInjectedByJsPlugin() ],
+  plugins: [ renderer(), vue() ],
   build: {
     outDir: 'dist',
     emptyOutDir: false,