+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   - Small update to content security policy
   - Dependencies bump
   - Removed stylus-loader (unnecessary)
+  - Added adjustable columns in the Job Manager
 
 ## v2.0.1
   - Removed `close` button from the input window to prevent accidental closure.