+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - Dependencies bump
   - Removed stylus-loader (unnecessary)
   - Added adjustable columns in the Job Manager
+  - Disabled devTools in production mode
 
 ## v2.0.1
   - Removed `close` button from the input window to prevent accidental closure.