+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - Removed stylus-loader (unnecessary)
   - Removed CSS injection plug-in for Vite
   - Added exception checking on AppSettings.getData
+  - Added a save check on settings window close
 
 ## v2.0.1
   - Removed `close` button from the input window to prevent accidental closure.